+++ gegelatilib/src/data/constant.cpp
@@ -38,12 +38,12 @@
 
 Data::Constant::operator int32_t() const
 {
-    return this->value;
+    return (int32_t)this->value;
 }
 
 Data::Constant::operator double() const
 {
-    return (double)this->value;
+    return this->value;
 }
 
 bool Data::Constant::operator==(const Constant& other) const
