+++ gegelatilib/src/learn/classificationLearningEnvironment.cpp
@@ -39,13 +39,13 @@
 
 #include "learn/classificationLearningEnvironment.h"
 
-void Learn::ClassificationLearningEnvironment::doAction(uint64_t actionID)
+void Learn::ClassificationLearningEnvironment::doAction(double actionID)
 {
     // Base method
     LearningEnvironment::doAction(actionID);
 
     // Classification table update
-    this->classificationTable.at(this->currentClass).at(actionID)++;
+    this->classificationTable.at(this->currentClass).at((uint64_t)actionID)++;
 }
 
 const std::vector<std::vector<uint64_t>>& Learn::
